Add existing rosterGenerator Script
This script is in a functional state, but needs some re-work

Moving to git repo so that it can be worked on remote and from
wherever is convenient. Also so that it can eventually be shared
with the Bayside IT department if they choose to incorporate the
functionality into their systems.